Issue #3079649 by Kuntyi: Step should be float

+++ js/range-slider-widget.js
@@ -5,11 +5,11 @@
         var $element = $(this);
         var min = parseInt($element.attr("min")) || 0;
         var max = parseInt($element.attr("max")) || (min+100);
-        var step = parseInt($element.attr("step")) || 0.1;
+        var step = parseFloat($element.attr("step")) || 0.1;
         var value = parseInt($element.val()) || min;
         var $container = $element.parent();
         var $slider = $container && $container.find('.range-slider-widget');
-        
+
         if($slider && $slider.length > 0) {
           $slider.slider({
             step: step,
@@ -21,9 +21,9 @@
             }
           });
         }
-        
+
       });
-      
+
     }
   };
 })(jQuery, window, Drupal);
